+++ src/pages/Search/SearchStyle.jsx
